+++ pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
   collection: ^1.19.1
   dio: ^5.8.0+1
   envied: ^1.1.1
+  extended_sliver: ^2.1.3
   flutter:
     sdk: flutter
   flutter_naver_map: ^1.3.0+1
